pyproject.toml
[project]
name = "zte-router-tool"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "requests",
    "python-dotenv",
]
